support openssl 1.1.1d: CipherSuites and CipherList
Add a CipherSuites SSL configuration option for TLSv13 ciphers.
CipherList is used only for TLSv12 ciphers.

OpenSSL allows CipherSuites to be empty - to ensure no TLSv13 ciphers
available.  This does not mean a client can connect to ngircd using
the TLSv12 ciphers which *are* enabled: instead *all* client
connections will fail by default!

This is because the so-called "Maximum Supported Version" is still
TLSv13 and the server drops the connection in the handshake with an
SSL_R_NO_CIPHERS_AVAILABLE "No ciphers enabled for max supported
SSL/TLS version" error. By default the max proto version is 0 which
means, the manpage says "automatically use the highest version
supported by the library."  But 0 really means "Only ever use the
TLSv13 - even if it is unavilable!

To handle this, if the CipherSuite is empty then we explicitly set
SSL_CTX_set_max_proto_version to TLSv12 - so clients can continue to
use Tlsv12 connections.